pyproject.toml
[project]
name = "gestao-codigo"
version = "0.2.0"
description = ""
authors = [
    {name = "Rodrigo Anjinho Gomes"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "duckdb (>=1.3.0,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "pydantic[email] (>=2.11.5,<3.0.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "alembic (>=1.16.1,<2.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "tzdata (>=2025.2,<2026.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]

[tool.taskipy.tasks]
format = "ruff format --preview ."
lint = "ruff check --fix --exit-non-zero-on-fix"
dev = "uvicorn webserver:app --port 8080 --host 0.0.0.0 --reload"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
taskipy = "^1.14.1"
ruff = "^0.11.12"
pytest = "^8.3.5"

